OffsetTables

Documentation for OffsetTables.

OffsetTables.OffsetTable

OffsetTables.sample

OffsetTables.sample

OffsetTables.OffsetTable — Type

OffsetTable{T<:Unsigned=UInt, I<:Integer=Int}(weights::AbstractVector{<:Real}; normalize=true)

An efficient data structure for sampling from a discrete distribution.

Maps every value representable by T to a value of type I in eachindex(wights) such that the number of values maped to a given index of weights is proportional to the value at that index.

The mapping can be accessed directly via OffsetTables.sample(x::T, ot::OffsetTable{T, I}) or indirectly by passing a random number generator which will be used to generate a random input of type T for you via OffsetTables.sample(rng::Random.AbstractRNG, ot::OffsetTable{T, I}) or simply via the Random API: rand(ot), rand(rng, ot), rand(ot, dims...), etc.

Set normalize = false for incrased performance when the weights are already normalized to sum to exactly the number of values representable by T (i.e. typemax(T)+1). A different sum will result in an error unless exactly one weight is non-zero, in which case the sum is not checked and the OffsetTable represents a constant distribution which always produces the index of the nonzero weight.

OffsetTables.sample — Method

sample(rng::Random.AbstractRNG, ot::OffsetTable{T, I}) -> I

Sample from ot using randomness drawn from rng.

Produces a random sample from the distribution represented by ot.

See also OffsetTable, Random.rand

OffsetTables.sample — Method

sample(x::T, ot::OffsetTable{T, I}) -> I

Sample from ot using the seed x.

If x is chosen uniformly at random from the set of all values representable by T then the output will be a random sample from the distribution represented by ot. The mapping is deterministic and not pseudo-random so for patterned input x the output will be patterned as well.

See also OffsetTable